chore(collaboratormodal): rename subfolder to components
